Went back to using $array, $dict, etc... for our macro shortcuts. Also, $dict and $mdict now use key/value order instead of value/key.

+++ OPExtensionKit/Source/Macros/OPMacros.h
@@ -41,7 +41,7 @@ void OPProfile(NSString *label, void(^task)(void));
  Returns the first non-nil object.
  */
 #define OPCoalesce(...)     OPCoalescePrivate(VA_ARGS_COUNT(__VA_ARGS__), VA_ARGS_C_ARRAY(__VA_ARGS__))
-id OPCoalescePrivate(int count, id __unsafe_unretained objs[]);
+id OPCoalescePrivate(NSUInteger count, id __unsafe_unretained objs[]);
 
 // Allows all orientations for the iPad, and everything but upside down portrait on the iPhone
 #define UI_ACCEPTABLE_INTERFACE_ORIENTATIONS(orientation)  (UI_USER_INTERFACE_IDIOM()==UIUserInterfaceIdiomPad ? YES : (UIInterfaceOrientationIsLandscape(orientation)||orientation==UIInterfaceOrientationPortrait))
